+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 - fixed `step_fwd/back()` for anything that wasn't `all=True/False/'loc'`
 - replacement of single statement no longer replaces that statement's `FST`, allowing `walk()` to recurse into replaced children
+- disallow put `_` as name for `MatchMapping.rest` and `MatchClass.kwd_attrs`
 
 
 ## 0.2.6 - alpha - 2026-01-31